+++ packages/components/.storybook/decorators/withProviders.tsx
@@ -1,4 +1,4 @@
-import React, { useEffect } from 'react'
+import React from 'react'
 
 import { LoadingState, ThemeProvider } from '@baseapp-frontend/design-system'
 import { RelayTestProvider, createTestEnvironment } from '@baseapp-frontend/graphql'
@@ -7,6 +7,7 @@ import { NotificationProvider } from '@baseapp-frontend/utils'
 import type { StoryContext, StoryFn } from '@storybook/react'
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query'
 
+import CurrentProfileProvider from '../../modules/profiles/context/CurrentProfileProvider'
 import '../../styles/tailwind/globals.css'
 import defaultTheme from '../__mocks__/theme'
 
@@ -15,18 +16,14 @@ const queryClient = new QueryClient()
 const withProviders = (Story: StoryFn, context: StoryContext) => {
   // TODO: registering a few tailwind classess (used by @baseapp-frontend/design-system components), need to figure out why the @baseapp-frontend/components storybook are not including it correctly
   // pb-3 px-3 w-full rounded-md bg-background-neutral px-2 py-1
-  const { environment, resolveMostRecentOperation } = createTestEnvironment()
+  const relayMockEnvironment = createTestEnvironment()
+  const { environment, queueOperationResolver } = relayMockEnvironment
+
+  context.parameters.relayMockEnvironment = relayMockEnvironment
 
   const mockResolvers = context.parameters.mockResolvers || {}
 
-  useEffect(() => {
-    try {
-      resolveMostRecentOperation({ mockResolvers })
-    } catch (error) {
-      // eslint-disable-next-line
-      console.log(error)
-    }
-  }, [mockResolvers, resolveMostRecentOperation])
+  queueOperationResolver(mockResolvers)
 
   return (
     <QueryClientProvider client={queryClient}>
@@ -34,7 +31,9 @@ const withProviders = (Story: StoryFn, context: StoryContext) => {
         <React.Suspense fallback={<LoadingState />}>
           <ThemeProvider {...defaultTheme}>
             <NotificationProvider>
-              <Story />
+              <CurrentProfileProvider>
+                <Story />
+              </CurrentProfileProvider>
             </NotificationProvider>
           </ThemeProvider>
         </React.Suspense>