+++ packages/components/__mocks__/styleMock.ts
@@ -1,0 +1,3 @@
+module.exports = require('@baseapp-frontend/test/__mocks__/styleMock.ts')
+
+export {}