+++ packages/components/modules/navigations/Header/AccountMenu/AccountPopover/LogoutItem/types.ts
@@ -1,0 +1,6 @@
+import type { PropsWithChildren } from 'react'
+
+export interface LogoutItemProps extends PropsWithChildren {
+  handlePopoverOnClose: () => void
+  logoutButtonLabel?: string
+}