+++ packages/components/modules/navigations/Header/AccountMenu/AccountPopover/MenuItems/types.ts
@@ -1,0 +1,6 @@
+import type { MenuItemProps } from '../../types'
+
+export interface MenuItemsProps {
+  handlePopoverOnClose: () => void
+  menuItems?: MenuItemProps[]
+}