+++ packages/components/modules/profiles/ProfilePopover/ProfilesList/LoadingState/styled.tsx
@@ -1,0 +1,7 @@
+import { Skeleton, styled } from '@mui/material'
+
+export const ProfileMenuItemSkeleton = styled(Skeleton)(({ theme }) => ({
+  width: '100%',
+  height: 52,
+  borderRadius: theme.spacing(0.75),
+}))