+++ packages/components/modules/profiles/context/CurrentProfileProvider/__tests__/__mock__/profiles.ts
@@ -1,0 +1,20 @@
+import { faker } from '@faker-js/faker'
+
+export const mockUserProfileFactory = (id: string) => {
+  return {
+    data: {
+      me: {
+        profile: {
+          id,
+          name: faker.person.fullName(),
+          image: {
+            url: faker.image.avatar(),
+          },
+          urlPath: {
+            path: faker.internet.url(),
+          },
+        },
+      },
+    },
+  }
+}