+++ packages/components/modules/profiles/graphql/queries/UserProfile.ts
@@ -1,0 +1,11 @@
+import { graphql } from 'react-relay'
+
+export const UserProfileQuery = graphql`
+  query UserProfileQuery {
+    me {
+      profile {
+        ...ProfileItemFragment
+      }
+    }
+  }
+`