+++ packages/components/modules/profiles/index.ts
@@ -1,0 +1,5 @@
+// Contexts & Providers
+export { default as useCurrentProfile } from './context/useCurrentProfile'
+
+// Components
+export * from './ProfilePopover'